Split IT into different classes
This way we gain:

* Classes dedicated for some use cases
* More readable, more maintainable
* Parallel execution of the suite

Local tests showed that we gain a faster execution from 03:37 to 02:09 with 77 tests ran by 26 suites instead of 4 suites.
