+++ integration-tests/src/test/java/fr/pilato/elasticsearch/crawler/fs/test/integration/FsCrawlerTestChecksumIT.java
@@ -1,0 +1,80 @@
+/*
+ * Licensed to David Pilato (the "Author") under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Author licenses this
+ * file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package fr.pilato.elasticsearch.crawler.fs.test.integration;
+
+import fr.pilato.elasticsearch.crawler.fs.beans.Doc;
+import fr.pilato.elasticsearch.crawler.fs.beans.File;
+import fr.pilato.elasticsearch.crawler.fs.settings.Fs;
+import org.elasticsearch.action.search.SearchRequest;
+import org.elasticsearch.action.search.SearchResponse;
+import org.elasticsearch.search.SearchHit;
+import org.junit.Test;
+
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+import static fr.pilato.elasticsearch.crawler.fs.framework.JsonUtil.extractFromPath;
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.is;
+import static org.junit.Assume.assumeNoException;
+
+/**
+ * Test checksum crawler settings
+ */
+public class FsCrawlerTestChecksumIT extends AbstractFsCrawlerITCase {
+
+    @Test
+    public void test_checksum_md5() throws Exception {
+        try {
+            MessageDigest.getInstance("MD5");
+        } catch (NoSuchAlgorithmException e) {
+            assumeNoException(e);
+        }
+
+        Fs fs = startCrawlerDefinition()
+                .setChecksum("MD5")
+                .build();
+        startCrawler(getCrawlerName(), fs, endCrawlerDefinition(getCrawlerName()), null);
+        SearchResponse searchResponse = countTestHelper(new SearchRequest(getCrawlerName()), 1L, null);
+        for (SearchHit hit : searchResponse.getHits().getHits()) {
+            Object checksum = extractFromPath(hit.getSourceAsMap(), Doc.FIELD_NAMES.FILE).get(File.FIELD_NAMES.CHECKSUM);
+            assertThat(checksum, is("caa71e1914ecbcf5ae4f46cf85de8648"));
+        }
+    }
+
+    @Test
+    public void test_checksum_sha1() throws Exception {
+        try {
+            MessageDigest.getInstance("SHA-1");
+        } catch (NoSuchAlgorithmException e) {
+            assumeNoException(e);
+        }
+
+        Fs fs = startCrawlerDefinition()
+                .setChecksum("SHA-1")
+                .build();
+        startCrawler(getCrawlerName(), fs, endCrawlerDefinition(getCrawlerName()), null);
+        SearchResponse searchResponse = countTestHelper(new SearchRequest(getCrawlerName()), 1L, null);
+        for (SearchHit hit : searchResponse.getHits().getHits()) {
+            Object checksum = extractFromPath(hit.getSourceAsMap(), Doc.FIELD_NAMES.FILE).get(File.FIELD_NAMES.CHECKSUM);
+            assertThat(checksum, is("81bf7dba781a1efbea6d9f2ad638ffe772ba4eab"));
+        }
+    }
+}
